document CORS on /embed/badges.html.
Didn't realize before that the only reason the embed script was
not running into cross-domain problems is because we are embedding
it on the same domain it's asking for stuff from. That will not be
normal. discovered problem by running the embed page from
www.impactstory.org/api-docs, instead of the same root domain. The
magic of CORS to the rescue, though.