When logged in /create and homepage both redirect to user's collection.
If the user has multiple collections, it just picks the one whose ID sorts first.
The user can still access other collections from the dropdown by clicking on
her name.

A consequence is that you must logout to view the homepage again. This is
like Mendeley, ResearchGate, and others...I think it's a feature, not a bug. I
reckon we've already sold 'em if they've made a profile...let's have user focus
on that.