Remove ability to add or remove products from collections.
They're deprecated now, so we don't want anyone investing time in 'em. We
could give a "you can't do this now" warning, but seemed more straightforward
to just remove the affordance. Some may be confused, so going to add a big
"this is deprecated now yo" banner to top as well.