First crack at redesigned front page.
Not 100% on the blue, but I kinda like it and imho good enough for now. Happy to discuss.
Note the radial gradient behind headline doesn't work in IE<10, but then nothing else we've
done is tested to work there, either. Is problem for later.

Still have to redo wording on the 3 supporting points below.