Send the user create AJAX call to the same domain the browser is using.
We were using the totalimpact.herokuapp domain name before, and the rest of the
application doesn't know about that cookie then. At some point, I think we can
get rid of the webapp_pretty url and just point everything to one webapp address. But
this would affect up the embed code now I think and so is out of scope.

Closes #46.