Items with no metrics can't zoom in, even when user clicks "expand all."
There's nothing really to zoom in on, so better just to not get all involved in what to say
there. Just don't zoom.

Note that the "n items with no metrics" list doesn't care whether
you've clicked "expand all" or not...it's expansion is controlled totally separately. There are
pros and cons to this (might be nice to expand it when user clicks expand all), but it was the
old behavior so I am leaving it alone.

That is to say, I am leaving it alone after tryingto change it and realizing it was going to be hard :)

Fixes #56.