Sets a tiPageType global for js using Flask template vars.
These are set with a dict in in views_helpers.py, to make the categories
easy to change and add to later if we want. The g.page_type global is injected
into all templates just before they are rendered.

The only potentially fragile thing about this is that it looks up page types based on
profile names, so if you change those it breaks. Had to decide whether to look up
based on 1) template names, 2) url paths, or 3) view function names. All had pros
and cons; happy to change (tho not sure how to do [3] actually).

Fixes #70, since we detect profiles base on use of the user-profile.html template.