Pageload reporting seems to be working with analytics.ready() callback.
Need to test that it's actually getting sent correctly, but it's not throwing
errors any more at least. Fixes #74, although more testing needed.

Also, realized the page_type code does not belong in Python...it's really
just an analytics issue that the client should be dealing with. tore out
the python that found analytics and replaced it with js. Much tidier.