Adds /users/test endpoint to delete test users.
Deletes three things: the user in the webapp postgres db, the
collection in the api couch db, and all the items that were in those
collections in the api couch db.  Need to be a bit careful with this
latter one, as items may be being used in other places, too.